[TD-888] 사장님 가게 오픈 정보 갱신시, 위치와 상관없이 오픈 정보가 유지되도록 수정 (AS-IS: 기존에는 반경 1KM를 벗어나면 강제 영업 종료)
