+++ threedollar-api-boss-service/src/main/kotlin/com/depromeet/threedollar/api/bossservice/service/store/BossStoreOpenService.kt
@@ -4,7 +4,6 @@ import java.time.LocalDateTime
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import com.depromeet.threedollar.api.core.service.bossservice.store.BossStoreServiceHelper
-import com.depromeet.threedollar.api.core.service.bossservice.store.dto.response.BossStoreOpenStatusResponse
 import com.depromeet.threedollar.common.exception.model.ForbiddenException
 import com.depromeet.threedollar.common.exception.type.ErrorCode
 import com.depromeet.threedollar.common.model.LocationValue
@@ -18,31 +17,23 @@ class BossStoreOpenService(
 ) {
 
     @Transactional
-    fun openBossStore(bossStoreId: String, bossId: String, mapLocation: LocationValue): BossStoreOpenStatusResponse {
+    fun openBossStore(bossStoreId: String, bossId: String, mapLocation: LocationValue) {
         val bossStore = BossStoreServiceHelper.findBossStoreByIdAndBossId(bossStoreRepository, bossStoreId = bossStoreId, bossId = bossId)
         bossStore.updateLocation(latitude = mapLocation.latitude, longitude = mapLocation.longitude)
         bossStoreRepository.save(bossStore)
 
-        val openDateTime = LocalDateTime.now()
-        bossStoreOpenTimeRepository.set(bossStoreId = bossStoreId, openDateTime = openDateTime)
-        return BossStoreOpenStatusResponse.open(openDateTime)
+        bossStoreOpenTimeRepository.set(bossStoreId = bossStoreId, openDateTime = LocalDateTime.now())
     }
 
     @Transactional
-    fun renewBossStoreOpenInfo(bossStoreId: String, bossId: String, mapLocation: LocationValue): BossStoreOpenStatusResponse {
+    fun renewBossStoreOpenInfo(bossStoreId: String, bossId: String, mapLocation: LocationValue) {
         val bossStore = BossStoreServiceHelper.findBossStoreByIdAndBossId(bossStoreRepository, bossStoreId = bossStoreId, bossId = bossId)
         validateIsOpenStore(bossStoreId = bossStoreId)
         if (bossStore.hasChangedLocation(latitude = mapLocation.latitude, longitude = mapLocation.longitude)) {
-            if (!bossStore.canBeRenewOpenDistance(mapLocation)) {
-                bossStoreOpenTimeRepository.delete(bossStoreId = bossStoreId)
-                return BossStoreOpenStatusResponse.close()
-            }
             bossStore.updateLocation(latitude = mapLocation.latitude, longitude = mapLocation.longitude)
             bossStoreRepository.save(bossStore)
         }
-
-        val openDateTime = upsertStoreOpenInfo(bossStoreId = bossStoreId)
-        return BossStoreOpenStatusResponse.open(openDateTime)
+        upsertStoreOpenInfo(bossStoreId = bossStoreId)
     }
 
     private fun validateIsOpenStore(bossStoreId: String) {
@@ -51,18 +42,16 @@ class BossStoreOpenService(
         }
     }
 
-    private fun upsertStoreOpenInfo(bossStoreId: String): LocalDateTime {
+    private fun upsertStoreOpenInfo(bossStoreId: String) {
         val openDateTime: LocalDateTime = bossStoreOpenTimeRepository.get(bossStoreId)
             ?: LocalDateTime.now()
         bossStoreOpenTimeRepository.set(bossStoreId = bossStoreId, openDateTime = openDateTime)
-        return openDateTime
     }
 
     @Transactional
-    fun closeBossStore(bossStoreId: String, bossId: String): BossStoreOpenStatusResponse {
+    fun closeBossStore(bossStoreId: String, bossId: String) {
         BossStoreServiceHelper.validateExistsBossStoreByBoss(bossStoreRepository, bossStoreId = bossStoreId, bossId = bossId)
         bossStoreOpenTimeRepository.delete(bossStoreId = bossStoreId)
-        return BossStoreOpenStatusResponse.close()
     }
 
 }
